+++ src/pages/public_registration/details-panel/_components/citation-box/_components/CopyCitationButton.tsx
